# api-gateway/docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "QuickServe API",
    "description": "API pour l'application QuickServe de transport et services",
    "version": "1.0.0",
    "contact": {
      "name": "Équipe QuickServe",
      "email": "support@quickserve.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Serveur de développement local"
    },
    {
      "url": "https://api.quickserve.com/api",
      "description": "Serveur de production"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Message d'erreur"
          }
        }
      },
...
      "Service": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "description": "Identifiant unique du service"
          },
          "client_id": {
            "type": "integer",
            "format": "int64",
            "description": "Identifiant du client"
          },
          "provider_id": {
            "type": "integer",
            "format": "int64",
            "description": "Identifiant du prestataire (null si pas encore assigné)"
          },
          "category_id": {
            "type": "integer",
            "format": "int64",
            "description": "Identifiant de la catégorie de service"
          },
          "category_name": {
            "type": "string",
            "description": "Nom de la catégorie de service"
          },
          "title": {
            "type": "string",
            "description": "Titre du service"
          },
          "description": {
            "type": "string",
            "description": "Description détaillée du service"
          },
          "address": {
            "type": "string",
            "description": "Adresse du service"
          },
          "latitude": {
            "type": "number",
            "format": "float",
            "description": "Latitude de l'adresse"
          },
          "longitude": {
            "type": "number",
            "format": "float",
            "description": "Longitude de l'adresse"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "assigned",
              "in_progress",
              "completed",
              "cancelled"
            ],
            "description": "Statut du service"
          },
          "payment_status": {
            "type": "string",
            "enum": [
              "pending",
              "completed",
              "failed",
              "cancelled"
            ],
            "description": "Statut du paiement"
          },
          "payment_amount": {
            "type": "number",
            "format": "float",
            "description": "Montant du paiement"
          },
          "scheduled_at": {
            "type": "string",
            "format": "date-time",
            "description": "Date planifiée pour le service"
          },
          "completed_at": {
            "type": "string",
            "format": "date-time",
            "description": "Date de complétion du service"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "Date de création de la demande de service"
          }
        }
      },
      "ServiceCreateRequest": {
        "type": "object",
        "required": [
          "category_id",
          "title",
          "address"
        ],
        "properties": {
          "category_id": {
            "type": "integer",
            "format": "int64",
            "description": "Identifiant de la catégorie de service"
          },
          "title": {
            "type": "string",
            "description": "Titre du service"
          },
          "description": {
            "type": "string",
            "description": "Description détaillée du service"
          },
          "address": {
            "type": "string",
            "description": "Adresse du service"
          },
          "latitude": {
            "type": "number",
            "format": "float",
            "description": "Latitude de l'adresse"
          },
          "longitude": {
            "type": "number",
            "format": "float",
            "description": "Longitude de l'adresse"
          },
          "scheduled_at": {
            "type": "string",
            "format": "date-time",
            "description": "Date planifiée pour le service"
          }
        }
      },
      "Rating": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "description": "Identifiant unique de l'évaluation"
          },
          "service_id": {
            "type": "integer",
            "format": "int64",
            "description": "Identifiant du service"
          },
          "rating": {
            "type": "integer",
            "minimum": 1,
            "maximum": 5,
            "description": "Note (de 1 à 5)"
          },
          "comment": {
            "type": "string",
            "description": "Commentaire"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "Date de création de l'évaluation"
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/services": {
      "post": {
        "tags": [
          "Services"
        ],
        "summary": "Créer un service",
        "description": "Créer une nouvelle demande de service",
        "operationId": "createService",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ServiceCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Service créé avec succès",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Prestation créée avec succès"
                    },
                    "service": {
                      "$ref": "#/components/schemas/Service"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Données invalides",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Non authentifié",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Catégorie non trouvée",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/services/{id}": {
      "get": {
        "tags": [
          "Services"
        ],
        "summary": "Détails d'un service",
        "description": "Récupérer les informations détaillées d'un service",
        "operationId": "getServiceDetails",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Identifiant du service",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Détails du service récupérés avec succès",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "service": {
                      "$ref": "#/components/schemas/Service"
                    },
                    "ratings": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Rating"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Non authentifié",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Accès non autorisé",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Service non trouvé",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/services/available": {
      "get": {
        "tags": [
          "Services"
        ],
        "summary": "Services disponibles",
        "description": "Récupérer la liste des services disponibles pour les prestataires",
        "operationId": "getAvailableServices",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Liste des services disponibles récupérée avec succès",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "services": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Service"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Non authentifié",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Accès non autorisé (rôle non prestataire)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/clients/services": {
      "get": {
        "tags": [
          "Services"
        ],
        "summary": "Services d'un client",
        "description": "Récupérer la liste des services demandés par le client connecté",
        "operationId": "getClientServices",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "description": "Filtre par statut de service",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "pending",
                "assigned",
                "in_progress",
                "completed",
                "cancelled"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Liste des services récupérée avec succès",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "services": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Service"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Non authentifié",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/providers/services": {
      "get": {
        "tags": [
          "Services"
        ],
        "summary": "Services d'un prestataire",
        "description": "Récupérer la liste des services assignés au prestataire connecté",
        "operationId": "getProviderServices",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "description": "Filtre par statut de service",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "assigned",
                "in_progress",
                "completed",
                "cancelled"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Liste des services récupérée avec succès",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "services": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Service"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Non authentifié",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/services/{id}/accept": {
      "post": {
        "tags": [
          "Services"
        ],
        "summary": "Accepter un service",
        "description": "Accepter un service disponible (pour prestataires)",
        "operationId": "acceptService",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Identifiant du service",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Service accepté avec succès",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Prestation acceptée avec succès"
                    },
                    "service": {
                      "$ref": "#/components/schemas/Service"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Service non disponible",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Non authentifié",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Accès non autorisé (rôle non prestataire)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/services/{id}/start": {
      "post": {
        "tags": [
          "Services"
        ],
        "summary": "Démarrer un service",
        "description": "Commencer l'exécution d'un service accepté (pour prestataires)",
        "operationId": "startService",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Identifiant du service",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Service démarré avec succès",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Prestation démarrée avec succès"
                    },
                    "service": {
                      "$ref": "#/components/schemas/Service"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Non authentifié",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Accès non autorisé ou service non assigné",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/services/{id}/complete": {
      "post": {
        "tags": [
          "Services"
        ],
        "summary": "Terminer un service",
        "description": "Marquer un service comme terminé (pour prestataires)",
        "operationId": "completeService",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Identifiant du service",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Service terminé avec succès",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Prestation terminée avec succès"
                    },
                    "service": {
                      "$ref": "#/components/schemas/Service"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Non authentifié",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Accès non autorisé ou service pas en cours",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/services/{id}/cancel": {
      "post": {
        "tags": [
          "Services"
        ],
        "summary": "Annuler un service",
        "description": "Annuler un service (pour clients ou prestataires)",
        "operationId": "cancelService",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Identifiant du service",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "reason": {
                    "type": "string",
                    "description": "Raison de l'annulation"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Service annulé avec succès",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Prestation annulée avec succès"
                    },
                    "service": {
                      "$ref": "#/components/schemas/Service"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Le service ne peut pas être annulé",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Non authentifié",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Accès non autorisé",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}